Refactor: Rename FilterName to helper
it makes more sense!!!
